ubuntu:22.04 add NOLINT to boost TEST macros allowing clang-tidy to find issues clang-tidy remove fuchsia-trailing-return seems buggy tests use constexpr Numbers avoiding magic number warnings fix 'const auto *', &'s suppress cppcoreguidelines-owning-memory in tests use upper case suffixes more [[nodiscard]] remove redundant TypeList template better macro argument names remove most 'using <namespace>' genhtml --demangle-cpp

+++ Tests/TestStructs.h
@@ -29,7 +29,7 @@ struct Struct
 struct User
 {
 	std::string Name;
-	int Age;
+	uint16_t Age;
 	std::list<std::string> Hobbies;
 };
